Fix tag() codegen for parser rules and split LALBlockCodegen into 3 classes
When a LAL rule has a parser (json/yaml/text), inputType from SPI is now
nulled out in GenCtx — it only applies to parser-less rules. This fixes
tag("KEY") generating null for network-profiling-slow-trace (MESH layer
with json{} parser), where SPI inputType was HTTPAccessLogEntry.

tag() now fails fast at compile time with clear error messages instead of
silently falling through. The generic fallback also throws instead of
generating null.

Split the 1600-line LALBlockCodegen into focused classes:
- LALBlockCodegen (725 lines): extractor/sink/sampler block scaffolding
- LALValueCodegen (903 lines): value access, conditions, data-source access
- LALDefCodegen (360 lines): def variable declarations and chain resolution

All methods documented with Javadoc and LAL script examples.
